refactor: move common fields to data layer
