+++ src/data-layer/collections/Semester.ts
@@ -1,4 +1,4 @@
-import { createTimeField } from '@/common/fields/date-time'
+import { createTimeField } from '@/data-layer/fields/DateTime'
 import { Weekday } from '@/types/types'
 import type { CollectionConfig } from 'payload'
